# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Attention Dragons API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /characters/{characterId}/journal:
    get:
      operationId: listJournalEntries
      tags: [journal]
      summary: List all journal entries for a character
      parameters:
        - name: characterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of journal entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JournalEntry"
    post:
      operationId: createJournalEntry
      tags: [journal]
      summary: Create a new journal entry
      parameters:
        - name: characterId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJournalEntryRequest"
      responses:
        "201":
          description: Journal entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
  /characters/{characterId}/journal/{entryId}:
    get:
      operationId: getJournalEntry
      tags: [journal]
      summary: Get a specific journal entry
      parameters:
        - name: characterId
          in: path
          required: true
          schema:
            type: integer
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Journal entry found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateJournalEntry
      tags: [journal]
      summary: Update a journal entry
      parameters:
        - name: characterId
          in: path
          required: true
          schema:
            type: integer
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJournalEntryRequest"
      responses:
        "200":
          description: Journal entry updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
    delete:
      operationId: deleteJournalEntry
      tags: [journal]
      summary: Delete a journal entry
      parameters:
        - name: characterId
          in: path
          required: true
          schema:
            type: integer
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Journal entry deleted
  /quests:
    get:
      operationId: listQuests
      tags: [quests]
      summary: List all quests
      parameters:
        - name: partyId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, completed, failed]
      responses:
        "200":
          description: List of quests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quest"
    post:
      operationId: createQuest
      tags: [quests]
      summary: Create a new quest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestRequest"
      responses:
        "201":
          description: Quest created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quest"
  /quests/{questId}:
    get:
      operationId: getQuest
      tags: [quests]
      summary: Get a specific quest
      parameters:
        - name: questId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quest found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quest"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateQuest
      tags: [quests]
      summary: Update a quest
      parameters:
        - name: questId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuestRequest"
      responses:
        "200":
          description: Quest updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quest"
    delete:
      operationId: deleteQuest
      tags: [quests]
      summary: Delete a quest
      parameters:
        - name: questId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Quest deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    JournalEntry:
      type: object
      properties:
        id:
          type: integer
        characterId:
          type: integer
        title:
          type: string
        body:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - characterId
        - title
        - body
        - createdAt
        - updatedAt
    CreateJournalEntryRequest:
      type: object
      properties:
        title:
          type: string
        body:
          type: string
      required:
        - title
        - body
    Quest:
      type: object
      properties:
        id:
          type: integer
        partyId:
          type: integer
          nullable: true
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, completed, failed]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - status
        - createdAt
        - updatedAt
    CreateQuestRequest:
      type: object
      properties:
        partyId:
          type: integer
          nullable: true
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, completed, failed]
      required:
        - title
        - description
    UpdateQuestRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, completed, failed]
        partyId:
          type: integer
          nullable: true