# lib/api-zod/src/generated/types/createQuestRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Attention Dragons API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateQuestRequestStatus } from "./createQuestRequestStatus";

export interface CreateQuestRequest {
  partyId?: number | null;
  title: string;
  description: string;
  status?: CreateQuestRequestStatus;
}