# lib/api-zod/src/generated/types/journalEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Attention Dragons API specification
 * OpenAPI spec version: 0.1.0
 */

export interface JournalEntry {
  id: number;
  characterId: number;
  title: string;
  body: string;
  createdAt: Date;
  updatedAt: Date;
}