# lib/api-zod/src/generated/types/quest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Attention Dragons API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuestStatus } from "./questStatus";

export interface Quest {
  id: number;
  partyId?: number | null;
  title: string;
  description: string;
  status: QuestStatus;
  createdAt: Date;
  updatedAt: Date;
}