# lib/api-zod/src/generated/types/updateQuestRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Attention Dragons API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateQuestRequestStatus } from "./updateQuestRequestStatus";

export interface UpdateQuestRequest {
  title?: string;
  description?: string;
  status?: UpdateQuestRequestStatus;
  partyId?: number | null;
}